不要なテストを削除
